chore(webpack): Use webpack5 configuration and check yarn build
