--- webpack/webpack.dev.js
@@ -1,7 +1,0 @@
-import merge from "webpack-merge";
-import common from "./webpack.common.js";
-
-export default merge(common, {
-  devtool: "inline-source-map",
-  mode: "development",
-});
