--- webpack/webpack.prod.js
@@ -1,6 +1,0 @@
-import merge from "webpack-merge";
-import common from "./webpack.common.js";
-
-export default merge(common, {
-  mode: "production",
-});
